# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Punjab Blood Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hospitals:
    get:
      operationId: searchHospitals
      tags: [public]
      summary: Search approved hospitals and blood banks
      parameters:
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: district
          schema: { type: string }
        - in: query
          name: bloodGroup
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: List of hospitals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HospitalPublic"
  /hospitals/{id}:
    get:
      operationId: getHospital
      tags: [public]
      summary: Get hospital details with blood inventory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Hospital details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HospitalDetail"
        "404":
          description: Not found
...
  /notifications/public:
    get:
      operationId: getPublicNotifications
      tags: [public]
      summary: Get published notifications visible to all users
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /admin/notifications:
    get:
      operationId: adminListNotifications
      tags: [admin]
      summary: List all notifications
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

    post:
      operationId: adminCreateNotification
      tags: [admin]
      summary: Create a notification or alert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /admin/notifications/{id}:
    patch:
      operationId: adminUpdateNotification
      tags: [admin]
      summary: Update a notification
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUpdate"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

    delete:
      operationId: adminDeleteNotification
      tags: [admin]
      summary: Delete a notification
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/notifications/{id}/publish:
    patch:
      operationId: adminPublishNotification
      tags: [admin]
      summary: Publish a notification (make visible to users)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # --- HOSPITAL: NOTIFICATIONS (blood donation camps) ---
  /hospital/notifications:
    get:
      operationId: hospitalListNotifications
      tags: [hospital]
      summary: List notifications submitted by the logged-in hospital
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

    post:
      operationId: hospitalCreateNotification
      tags: [hospital]
      summary: Submit a blood donation camp notification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalNotificationInput"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /hospital/notifications/{id}:
    delete:
      operationId: hospitalDeleteNotification
      tags: [hospital]
      summary: Retract a notification submitted by the logged-in hospital
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # --- ADMIN: CATEGORIES ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BloodInventoryItem:
      type: object
      required: [id, hospitalId, bloodGroup, units, status, lastUpdated]
      properties:
        id: { type: integer }
        hospitalId: { type: integer }
        bloodGroup: { type: string }
        units: { type: integer }
        status: { type: string }
        lastUpdated: { type: string }
...
    Notification:
      type: object
      required: [id, title, message, type, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        message: { type: string }
        type: { type: string }
        status: { type: string }
        hospitalId: { type: ["integer", "null"] }
        hospitalName: { type: ["string", "null"] }
        campDate: { type: ["string", "null"] }
        campLocation: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
        createdAt: { type: string }
    NotificationInput:
      type: object
      required: [title, message, type]
      properties:
        title: { type: string }
        message: { type: string }
        type: { type: string }
        expiresAt: { type: string }
    NotificationUpdate:
      type: object
      properties:
        title: { type: string }
        message: { type: string }
        type: { type: string }
        status: { type: string }
        expiresAt: { type: string }
    HospitalNotificationInput:
      type: object
      required: [title, message, campDate, campLocation]
      properties:
        title: { type: string }
        message: { type: string }
        campDate: { type: string }
        campLocation: { type: string }
        expiresAt: { type: string }
        # If true, the camp notification is published immediately. If false
        # (or omitted), it's submitted as pending_review for admin approval.
        publishNow: { type: boolean }

    # --- CATEGORIES ---
...
    HospitalPublic:
      type: object
      required: [id, name, type, ownership, city, district, address, contactNumber, status]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        ownership: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        pinCode: { type: ["string", "null"] }
        contactNumber: { type: string }
        email: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        googleMapsLat: { type: ["number", "null"] }
        googleMapsLng: { type: ["number", "null"] }
        workingHours: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        status: { type: string }
        bloodInventory:
          type: array
          items:
            $ref: "#/components/schemas/BloodInventoryItem"
    HospitalDetail:
      type: object
      required: [id, name, type, ownership, city, district, address, contactNumber, status, bloodInventory]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        ownership: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        pinCode: { type: ["string", "null"] }
        contactNumber: { type: string }
        email: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        googleMapsLat: { type: ["number", "null"] }
        googleMapsLng: { type: ["number", "null"] }
        workingHours: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        status: { type: string }
        bloodInventory:
          type: array
          items:
            $ref: "#/components/schemas/BloodInventoryItem"
...