# lib/api-zod/src/generated/types/hospitalNotificationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Punjab Blood Connect API
 * OpenAPI spec version: 0.2.0
 */

export interface HospitalNotificationInput {
  title: string;
  message: string;
  campDate: string;
  campLocation: string;
  expiresAt?: string;
  publishNow?: boolean;
}