# lib/api-zod/src/generated/types/hospitalPublic.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Punjab Blood Connect API
 * OpenAPI spec version: 0.2.0
 */
import type { BloodInventoryItem } from './bloodInventoryItem';

export interface HospitalPublic {
  id: number;
  name: string;
  type: string;
  ownership: string;
  city: string;
  district: string;
  address: string;
  /** @nullable */
  pinCode?: string | null;
  contactNumber: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  googleMapsLat?: number | null;
  /** @nullable */
  googleMapsLng?: number | null;
  /** @nullable */
  workingHours?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  status: string;
  bloodInventory?: BloodInventoryItem[];
}